basic test for separable KRILL
