+++ test/drivers/test_krillx.cc
@@ -16,6 +16,7 @@ using RandBLAS::DenseDist;
 using RandBLAS::SparseDist;
 using RandBLAS::RNGState;
 using RandLAPACK::linops::RegExplicitSymLinOp;
+using RandLAPACK::linops::SEKLO;
 using RandLAPACK_Testing::polynomial_decay_psd;
 
 
@@ -168,6 +169,21 @@ TEST_F(TestKrillx, test_krill_separable_rpchol) {
     for (int64_t decay = 2; decay < 4; ++decay) {
         auto G = polynomial_decay_psd(m, 1e12, (T) decay, 99);
         RegExplicitSymLinOp G_linop(m, G.data(), m, mus);
+        int64_t k = 128;
+        run_krill_separable(0, G_linop, k);
+        run_krill_separable(1, G_linop, k);
+    }
+}
+
+TEST_F(TestKrillx, test_krill_separable_squared_exp_kernel) {
+    using T = double;
+    T mu_min = 1e-2;
+    vector<T> mus {mu_min, mu_min*10, mu_min*100};
+    for (uint32_t key = 0; key < 5; ++key) {
+        //auto G = polynomial_decay_psd(m, 1e12, (T) decay, key);
+        //RegExplicitSymLinOp G_linop(m, G.data(), m, mus);
+        vector<T> X0 = RandLAPACK_Testing::random_gaussian_mat<T>(5, m, key);
+        SEKLO G_linop(m, X0.data(), 5, 3.0, mus);
         int64_t k = 128;
         run_krill_separable(0, G_linop, k);
         run_krill_separable(1, G_linop, k);
